+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add a window to modify movie settings
 * Add more extra rows in input editor
 * Add hex viewer
+* Open and seek hex viewer from ram watch
 
 ### Changed